Biographyのイベント履歴部分の２重管理を解消
