--- src/entities/biography/Collaboration.ts
@@ -1,16 +1,0 @@
-export type CollaborationLink = {
-    name: string;
-    url: string;
-};
-
-export type CollaborationProduct = {
-    name: string;
-    artist: string;
-}
-
-export type Collaboration = {
-    date: Date;
-    product: CollaborationProduct;
-    partOfTheWork: string;
-    links: CollaborationLink[];
-};
