--- src/entities/biography/EventHistory.ts
@@ -1,10 +1,0 @@
-export type EventLink = {
-    name: string;
-    url: string;
-};
-
-export type EventHistory = {
-    date: Date;
-    name: string;
-    links?: EventLink[];
-};
