--- src/entities/biography/Profile.ts
@@ -1,6 +1,0 @@
-export type Profile = {
-    name: string;
-    nameCaption?: string;
-    introductions: string[];
-    activities: string[];
-};
