+++ src/pages/biography.tsx
@@ -4,10 +4,10 @@ import PageHeader from '../components/common/PageHeader';
 import PageMeta from '../components/common/PageMeta';
 
 import Index, {Props as BiographyProps} from '../components/biography/Index';
-import { getProfile, listEventHistories, listCollaborations } from '../services/biography/BiographyService';
 import { getBiographyMeta } from '../services/common/MetaDataService';
 import { SupportedLocale } from '../constants/i18n';
 import { ensureSupportedLocale } from '../utilities/i18n';
+import { InMemoryBiographyService } from '../services/biography/InMemoryBiographyService';
 
 
 interface GetStaticProps {
@@ -16,14 +16,15 @@ interface GetStaticProps {
 
 export const getStaticProps = async ({locale}: GetStaticProps) => {
     const validatedLocale = ensureSupportedLocale(locale);
+    const biographyService = new InMemoryBiographyService();
     return {
         props: {
             locale: validatedLocale as SupportedLocale,
             meta: getBiographyMeta(validatedLocale),
             biographyProps: {
-                profile: getProfile(validatedLocale),
-                eventHistories: listEventHistories(validatedLocale),
-                collaborations: listCollaborations(validatedLocale),
+                profile: biographyService.getProfile(validatedLocale),
+                eventHistories: biographyService.listEventHistories(validatedLocale),
+                collaborations: biographyService.listCollaborations(validatedLocale),
             } as BiographyProps,
         }
     };
