+++ src/utilities/i18n.test.ts
@@ -1,8 +1,12 @@
+import { SupportedLocale } from "../constants/i18n";
 import { ensureSupportedLocale } from "./i18n";
 
-test('ensureSupportedLocale', () => {
-    expect(ensureSupportedLocale("en")).toBe("en");
-    expect(ensureSupportedLocale("ja")).toBe("ja");
-    // 非対応の場合はデフォルトに
-    expect(ensureSupportedLocale("zh")).toBe("ja");
+describe('ensureSupportedLocale', () => {
+    test('対応しているロケールはそのまま', () => {
+        expect(ensureSupportedLocale("en")).toBe("en");
+        expect(ensureSupportedLocale("ja")).toBe("ja");  
+    });
+    test('非対応のロケールはデフォルトに', () => {
+        expect(ensureSupportedLocale("zh")).toBe("ja");
+    });
 });
